backend/requirements.txt
fastapi
uvicorn
pymongo
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv